Ran full code cleanup, fix iotemp

+++ Lyo.Net/Core/Common/Lyo.Common/Security/CryptographicRandom.cs
@@ -2,14 +2,18 @@ using System.Security.Cryptography;
 
 namespace Lyo.Common.Security;
 
-/// <summary>Cryptographically secure random helpers compatible with netstandard2.0 (no dependency on <see cref="RandomNumberGenerator.GetBytes(int)"/> / <see cref="RandomNumberGenerator.GetInt32(int, int)"/>).</summary>
+/// <summary>
+/// Cryptographically secure random helpers compatible with netstandard2.0 (no dependency on <see cref="RandomNumberGenerator.GetBytes(int)" /> /
+/// <see cref="RandomNumberGenerator.GetInt32(int, int)" />).
+/// </summary>
 public static class CryptographicRandom
 {
-    /// <summary>Returns an array of <paramref name="length"/> cryptographically strong random bytes.</summary>
+    /// <summary>Returns an array of <paramref name="length" /> cryptographically strong random bytes.</summary>
     public static byte[] GetBytes(int length)
     {
         if (length < 0)
             throw new ArgumentOutOfRangeException(nameof(length));
+
         if (length == 0)
             return Array.Empty<byte>();
 
@@ -18,7 +22,7 @@ public static class CryptographicRandom
         return bytes;
     }
 
-    /// <summary>Fills <paramref name="buffer"/> with cryptographically strong random bytes.</summary>
+    /// <summary>Fills <paramref name="buffer" /> with cryptographically strong random bytes.</summary>
     public static void Fill(Span<byte> buffer)
     {
         if (buffer.IsEmpty)
@@ -35,7 +39,7 @@ public static class CryptographicRandom
 #endif
     }
 
-    /// <summary>Returns a random integer in <c>[<paramref name="fromInclusive"/>, <paramref name="toExclusive"/>)</c> with an approximately uniform distribution.</summary>
+    /// <summary>Returns a random integer in <c>[<paramref name="fromInclusive" />, <paramref name="toExclusive" />)</c> with an approximately uniform distribution.</summary>
     public static int GetInt32(int fromInclusive, int toExclusive)
     {
         if (fromInclusive >= toExclusive)
@@ -46,7 +50,7 @@ public static class CryptographicRandom
             return fromInclusive;
 
         var r = NextUInt64Exclusive(range);
-        return (int)((long)fromInclusive + (long)r);
+        return (int)(fromInclusive + (long)r);
     }
 
     private static ulong NextUInt64Exclusive(ulong range)
